+++ lib/aspera/preview/terminal.rb
@@ -44,7 +44,7 @@ module Aspera
           fit_term_ratio = [term_rows.to_f * font_ratio / image.rows.to_f, term_columns.to_f / image.columns.to_f].min
           height_ratio = double ? 2.0 : 1.0
           image = image.scale((image.columns * fit_term_ratio).to_i, (image.rows * fit_term_ratio * height_ratio / font_ratio).to_i)
-          # quantum depth is 8 or 16, see: `convert xc: -format "%q" info:`
+          # quantum depth is 8 or 16, see: `magick xc: -format "%q" info:`
           shift_for_8_bit = Magick::MAGICKCORE_QUANTUM_DEPTH - 8
           # get all pixel colors, adjusted for Rainbow
           pixel_colors = []