+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * `echo`: Display of list was showing only first element
     * `transferd`: support for version 1.1.5+
   * Breaking Changes:
+    * `preview`: Updated Image Magick to v7+
     * `aoc`: `admin subscription` split into `admin subscription account` and `admin subscription usage`
     * **agent**: `alpha` renamed to `desktop`